Fix GameClient::checkServerInput not iterating inputs properly
We were deleting the input after finding it by inputId if size > 2, which was a leftover from previous code I was too stupid to remove. This was completely useless (since ensuring that size > 2 is already done within the previous loop) and was probably causing some bugs.

Also, the iterator was not being reset before correcting abilities, which would cause bugs if a single input required to be corrected both in movement and abilities. This is *probably* what was causing the infamous ultimate recharge bug.